FIX delete picture

+++ functions/deletepicture.php
@@ -1,19 +1,14 @@
 <?php
    session_start();
-   if (isset($_SESSION['id'])){
+   if (isset($_SESSION['id']) && isset($_POST['id'])){
        require_once '../config/database.php';
-       $id = $_SESSION['id'];
-       $pdo = new PDO($DB_DSN, $DB_USER, $DB_PASSWORD);
-       $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-       $req = $pdo->prepare("SELECT * FROM gallery WHERE userid=:userid");
-       $req->execute(array(':userid' => $id));
-       $pic = $req->fetch();
-       $req->closeCursor();
-
+       $userid = $_SESSION['id'];
+       $imgid = $_POST['id'];
        echo unlink("../" . $pic['img']);
-       if ($_SESSION['id'] == $pic['userid']) {
-            $req = $pdo->prepare("DELETE FROM gallery WHERE id=:id");
-            $req->execute(array(':id' => $pic['id']));
-       }
+        $pdo = new PDO($DB_DSN, $DB_USER, $DB_PASSWORD);
+        $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+        $req = $pdo->prepare("DELETE FROM gallery WHERE id=:id AND userid=:userid");
+        $req->execute(array(':id' => $imgid, ':userid' => $userid));
+        $req->closeCursor();
    }
 ?>
